fix(setup): route to signup from ?step=signup, not a stale campaign cookie
Replaces the earlier "Log In" CTA approach (a UI band-aid) with a routing fix,
and reverts the LoginLink component + Signup/JoinWaitlist changes from the
first commit. Net change over origin/main is one file: src/app/(setup)/setup/page.tsx.

Root cause: PR #2346 widened /setup's skip-invite-gate to fire on a campaignTag
cookie. That cookie is a session cookie (no expiry) written on any
/invite?campaign= claim and cleared only on a successful new signup. A
returning user who claimed a campaign badge earlier in the same browser
session was routed past Landing (the only screen with a Log In button)
straight onto Signup, unable to log back in — and signing up would mint a
duplicate account colliding with their existing passkey.

/setup now decides "skip to signup" from the URL (?step=signup) — the signal
every campaign/skip flow already sends when it pushes to /setup
(ShhhhhLandingPage, InvitesPage.handleClaim) — plus the existing invite-code
cookie. It no longer trusts the persistent campaignTag cookie for routing.
useZeroDev still reads that cookie post-signup to award the badge, so #2346's
campaign/offramp arrivals are unaffected. A stale cookie on a bare /setup now
lands on Landing with Log In intact.
